requirements.txt
tkinterdnd2==0.4.2
pillow==10.3.0
requests==2.31.0
beautifulsoup4==4.12.3